[clean]appフォルダを見直し、componentsフォルダ等に機能ごとにファイルを分ける
