+++ app/createRecruitment/page.tsx
@@ -1,8 +1,8 @@
 import styles from "./createRecruitment.module.css";
 import { createClient } from "@/utils/supabase/server";
-import { fetchProfile } from "../supabase_function/profile";
+import { fetchProfile } from "../../lib/supabase_function/profile";
 import { redirect } from "next/navigation";
-import RecruitmentForm from "./form";
+import RecruitmentForm from "../../components/recruitment/create/form";
 
 const createRecruitment = async () => {
   const supabase = await createClient();
