+++ utils/enum/enum.tsx
@@ -1,3 +1,3 @@
-import { Constants } from "@/database.types";
+import { Constants } from "@/utils/supabase/database.types";
 
 export const STATUS_VALUES = Constants.public.Enums.status_enum;
